next step towards simplifying mouse handling code
The mouseMove handler needs to know how far the mouse was moved. Formerly, this had to be calculated by saving the last known value and comparing it to the new value. Now, browsers do this for us and give the value as MouseEvent.movementX (this is the Pointer Lock API)