Towards allowing non-contiguous and possibly overlapping segments
For kommunikationsart (formerly kwplus), I definitely need
non-contiguous segments. Overlapping segments would be nice to have, but
are definitely harder to get right.

This part of the change is non-invasive. It adds two functions to the
“libary api” – that is, the collection of static functions on
LevelService – and extends the result type of a third function.

The biggest problem in the new functions that I am aware of is that they
allow the segments to get out of order: A segment level’s items can now
be [itemA, itemB], with itemB having an earlier sampleStart than itemA.
This fundamentally violates an assumption of the function
getClosestItem.